# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  polling-system_auth_models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  polling-system_polling_models.Answer:
    properties:
      answer:
        type: string
      candidate:
        $ref: '#/definitions/polling-system_polling_models.Candidate'
      candidate_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      question:
        $ref: '#/definitions/polling-system_polling_models.PollQuestion'
      question_id:
        type: string
      source:
        type: string
    type: object
  polling-system_polling_models.AnswerCreate:
    properties:
      answer:
        type: string
      candidate_id:
        type: string
      source:
        type: string
    required:
    - answer
    - candidate_id
    type: object
...
  polling-system_polling_models.Candidate:
    properties:
      answers:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      poll:
        $ref: '#/definitions/polling-system_polling_models.Poll'
      poll_id:
        type: string
      updated_at:
        type: string
      voice_poll_sentiment:
        type: string
      voice_poll_started_at:
        type: string
      voicemail_detected:
        type: boolean
    type: object
...
  polling-system_polling_models.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
  polling-system_polling_models.PaginatedAnswers:
    properties:
      items:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  polling-system_polling_models.Poll:
    properties:
      bland_pathway_ids:
        type: object
      candidates:
        items:
          $ref: '#/definitions/polling-system_polling_models.Candidate'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/polling-system_auth_models.User'
      custom_voice_poll_instructions:
        type: string
      dont_send_sms:
        type: boolean
      first_typeform_sms_text:
        type: string
      id:
        type: string
      only_day_hours:
        type: boolean
      poll_name:
        type: string
      poller_name:
        type: string
      questions:
        items:
          $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        type: array
      scheduled_to_start_at:
        type: string
      status:
        type: string
      time_zone:
        type: string
      typeform_forms:
        type: object
      updated_at:
        type: string
      voice_poll_delay_minutes:
        type: integer
      voice_poll_method:
        type: string
    type: object
...
  polling-system_polling_models.PollQuestion:
    properties:
      answer_options:
        type: object
      answer_type:
        type: string
      answers:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      created_at:
        type: string
      default_next_question:
        $ref: '#/definitions/polling-system_polling_models.PollQuestion'
      default_next_question_id:
        type: string
      flow_rules:
        type: object
      id:
        type: string
      poll:
        $ref: '#/definitions/polling-system_polling_models.Poll'
      poll_id:
        type: string
      question:
        type: string
      randomize_answers:
        type: boolean
      typeform_field_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Unified API for authentication, polls, contacts, and file management.
  title: Polling System API
  version: "1.0"
paths:
  /answers/{id}:
    delete:
      description: Deletes an answer. Admin only.
      parameters:
      - description: Answer UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an answer
      tags:
      - Answers
    get:
      description: Returns a single answer.
      parameters:
      - description: Answer UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.Answer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get answer by ID
      tags:
      - Answers
...
  /candidates/{candidateId}/answers:
    get:
      description: Returns a paginated list of all answers submitted by a candidate.
      parameters:
      - description: Candidate UUID
        in: path
        name: candidateId
        required: true
        type: string
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PaginatedAnswers'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List answers by a candidate
      tags:
      - Answers
...
  /questions/{questionId}/answers:
    get:
      description: Returns a paginated list of answers for the given question.
      parameters:
      - description: Question UUID
        in: path
        name: questionId
        required: true
        type: string
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PaginatedAnswers'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List answers for a question
      tags:
      - Answers
    post:
      consumes:
      - application/json
      description: Submits an answer from a candidate. Each candidate can answer a
        question only once.
      parameters:
      - description: Question UUID
        in: path
        name: questionId
        required: true
        type: string
      - description: Answer to submit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.AnswerCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_polling_models.Answer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit answer for a question
      tags:
      - Answers
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"